+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html
 - Non-backward compatible changes:
     - Hexital and Indicator 'append' will now only append to Candles and not sort placement.
     - Renamed Indicator 'as_list' to 'readings'
+    - Removed 'has_reading' as pointless with method 'exists'
 
 - Reworked and optimised 'append' and added 'insert' and 'prepend'
 - Added IndicatorCollection as a way to better reference Indicator's